Route all fund accountability through the PL and close the audit's finance seams
The 2026-08-12 budget-automation audit, fixed end to end, plus the mandated
post-disbursement lifecycle:

- one fundable-lines law (apps/fund_requests/fundable.py) selects lines for
  every funding channel; approval routes advances to the accountant and the
  period channel no longer releases unconfirmed money
- per-line disbursement requires membership in an approved weekly request;
  returned requests rebuild from the corrected schedule instead of dead-ending
- snapshot-lock guards run row-locked inside the transaction; PL return of a
  monthly plan respects the accountant-action lock
- disbursed weeks require the owner's receipt confirmation before
  accountability opens; all accountability (advance and self-funded
  reimbursement claims) is approved by the PL before the Accountant, and IA
  verification gates every settlement including reimbursements
- disbursed/accounted rollups read the ledger so cancelled-after-payment cash
  stays visible; new money_moved_on_cancelled_work critical health check
- est_cost_cents widened to 64-bit; undated work is never priced; dead
  costing code removed; finance_status speaks the real weekly vocabulary

Five E2E tests pin the whole chain: catalogue-priced scheduling populates the
budget, the week compiles/routes/approves/disburses as one, receipt opens
accountability, the PL and IA gates hold, and the week closes only when every
line is settled.

+++ apps/planning/models.py
@@ -78,7 +78,7 @@ class MonthlyPlanActivity(TimeStampedModel):
     assignee_id = models.CharField(max_length=30, null=True, blank=True)
     # Despite the field name, this holds plain integer UGX (whole
     # shillings), not cents -- see apps.activities.models.Activity.est_cost_cents.
-    est_cost_cents = models.IntegerField(default=0)
+    est_cost_cents = models.BigIntegerField(default=0)
     status = models.CharField(max_length=32, default="Planned")
     intervention_area = models.CharField(max_length=64, null=True, blank=True)
     delivery_type = models.CharField(max_length=16, null=True, blank=True)
